+++ __init__.py
@@ -7,6 +7,8 @@ from .api.deploy import deploy_namespace
 from .models.challenge_type import ZyncChallengeType
 from .views.admin_config import define_zync_admin
 
+__version__ = "0.2.0"
+
 
 def load(app):
     app.db.create_all()